Extract markdown H1
