+++ src/docfx/models/PageMetadata.cs
@@ -1,14 +1,16 @@
 // Copyright (c) Microsoft. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using Newtonsoft.Json;
 using Newtonsoft.Json.Linq;
 
 namespace Microsoft.Docs.Build
 {
-    internal class PageModel
+    internal class PageMetadata
     {
-        public string Content { get; set; }
+        public string Title { get; set; }
 
-        public PageMetadata Metadata { get; set; }
+        [JsonExtensionData]
+        public JObject Metadata { get; set; }
     }
 }
